Change type of sky ratio so it is actually a ratio
Although all these values should probably be floats (#5), the checks
of this value against sky_threshold certainly weren't working as
designed with it being an int.

Fixes #4.

+++ lim-yen-wu.c
@@ -59,7 +59,7 @@ struct block_metrics
 	float brightness;
 	float mean_saturation;
 	float mean_hue;
-	int sky_ratio;
+	float sky_ratio;
 };
 
 struct figures_of_merit
